+++ packages/test-utils/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- set non-w3c mode for appium capabilities
 
 ## 1.0.3 - 2021/5/24